# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Creatools - TikTok LIVE monitoring API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /media:
    get:
      operationId: listMedia
      tags: [media]
      summary: List user media items
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of media items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaListResponse"
        "401":
          description: Unauthorized
  /media/upload:
    post:
      operationId: uploadMedia
      tags: [media]
      summary: Upload a media file via multipart/form-data (PNG, JPG, GIF, WebP, max 50 MB)
      description: "The 'file' field carries the binary image. MIME type and file type are verified server-side via magic bytes. Plan quota is enforced server-side."
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/MediaUploadInput"
      responses:
        "200":
          description: Uploaded item metadata and public URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaUploadResponse"
        "400":
          description: Invalid or missing file
        "401":
          description: Unauthorized
        "413":
          description: Storage quota exceeded for user plan
  /media/storage:
    get:
      operationId: getMediaStorage
      tags: [media]
      summary: Get storage usage (plan limit enforced server-side)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Storage usage and limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaStorageResponse"
        "401":
          description: Unauthorized
  /media/{id}:
    patch:
      operationId: patchMedia
      tags: [media]
      summary: Rename or change category of a media item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaPatchInput"
      responses:
        "200":
          description: Updated media item
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: "#/components/schemas/MediaItem"
                required: [item]
        "401":
          description: Unauthorized
        "404":
          description: Item not found
    delete:
      operationId: deleteMedia
      tags: [media]
      summary: Delete a media item and its file
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
        "401":
          description: Unauthorized
        "404":
          description: Item not found
...
  /layouts:
    get:
      operationId: getLayouts
      tags: [layouts]
      summary: List all saved layout presets for the authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of presets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LayoutPreset"
    post:
      operationId: createLayout
      tags: [layouts]
      summary: Save a new layout preset
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LayoutInput"
      responses:
        "201":
          description: Created preset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LayoutPreset"
        "400":
          description: Validation error or preset limit reached
  /layouts/{id}:
    patch:
      operationId: renameLayout
      tags: [layouts]
      summary: Rename a layout preset
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LayoutRenameInput"
      responses:
        "200":
          description: Updated preset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LayoutPreset"
        "404":
          description: Preset not found
    delete:
      operationId: deleteLayout
      tags: [layouts]
      summary: Delete a layout preset
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Preset not found
components:
  schemas:
...
    LayoutInput:
      type: object
      properties:
        name:
          type: string
        layers:
          type: array
          items:
            type: object
      required: [name, layers]
    LayoutRenameInput:
      type: object
      properties:
        name:
          type: string
      required: [name]
    LayoutPreset:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        layers:
          type: array
          items:
            type: object
        createdAt:
          type: number
        updatedAt:
          type: number
      required: [id, name, layers, createdAt, updatedAt]
...
    MediaUploadInput:
      type: object
      description: "Multipart upload fields. The binary 'file' field is sent as a separate form part and is not typed here to avoid browser-only File/Blob types in server-side schemas."
      properties:
        category:
          type: string
          enum: [Geral, Banners, Logos, QR Codes, Thumbnails]
          default: Geral
        name:
          type: string
          description: Optional display name; defaults to original filename
    MediaPatchInput:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [Geral, Banners, Logos, QR Codes, Thumbnails]
    MediaItem:
      type: object
      properties:
        id:
          type: string
        filename:
          type: string
        originalName:
          type: string
        category:
          type: string
        size:
          type: integer
          description: File size in bytes
        mimeType:
          type: string
        width:
          type: integer
          nullable: true
          description: Image width in pixels (null if not extractable)
        height:
          type: integer
          nullable: true
          description: Image height in pixels (null if not extractable)
        createdAt:
          type: string
          format: date-time
      required: [id, filename, originalName, category, size, mimeType, width, height, createdAt]
    MediaListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MediaItem"
      required: [items]
    MediaUploadResponse:
      type: object
      properties:
        item:
          $ref: "#/components/schemas/MediaItem"
        url:
          type: string
          description: Public URL path for the uploaded file
      required: [item, url]
    MediaStorageResponse:
      type: object
      properties:
        used:
          type: integer
          description: Bytes used
        limit:
          type: integer
          description: Bytes allowed by the user plan
        plan:
          type: string
          description: User plan (free, basic, pro)
        count:
          type: integer
          description: Number of files stored
        categories:
          type: array
          items:
            type: string
      required: [used, limit, plan, count, categories]